Add Istio sidecar enabled tests and refactor code
